fix: Update GCP Docs (#822)
## Description

Update self-managed installation for GCP

## Ticket

Does this PR fix an existing issue? If yes, provide a link to the ticket
here:

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [x] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [x] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [x] I merged the latest changes from `main` into my feature branch
before submitting this PR.

---------

Co-authored-by: Noah Luna <15202580+ngrayluna@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>