Recent updates to registry landing page (#827)
- update latest server release version to reflect bug fixes
- link to tags
- link to model reg end of life docs
- update details on legacy model reg

## Description

What does the pull request do? If it fixes a bug or resolves a feature
request, be sure to link to that issue.

## Ticket

Does this PR fix an existing issue? If yes, provide a link to the ticket
here:

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [ ] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [ ] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [ ] I merged the latest changes from `main` into my feature branch
before submitting this PR.

---------

Co-authored-by: Noah Luna <15202580+ngrayluna@users.noreply.github.com>