update media documentation with command for installing additional dependencies (#853)
## Description

This PR adds additional documentation to the media logging page on how
to install additional dependencies.

## Ticket

Does this PR fix an existing issue? If yes, provide a link to the ticket
here: N/A

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [x] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [x] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [x] I merged the latest changes from `main` into my feature branch
before submitting this PR.