Change titles of some platform pages to align with the positioning (#852)
## Description

Changed the titles of some platform pages to align better with how we
position those features or capabilities to customers. Some pages are
informational only and not necessarily action-oriented so we need to
label them accordingly.

## Ticket

NA

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.