Add information for project level roles to manage project access control (#856)
## Description

Adding the required overview, configuration instructions and other
details for the project level roles which is a new IAM feature.

## Ticket

NA

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.

---------

Co-authored-by: Noah Luna <15202580+ngrayluna@users.noreply.github.com>