Adding new supported regions for AWS and Azure (#861)
## Description

Adds newly available regions for Dedicated Deployments in AWS and Azure.
Specifically the ap-south-1 region in AWS and the Canada Central region
in Azure.

## Ticket

Does this PR fix an existing issue? If yes, provide a link to the ticket
here:

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.