Adding a caution that we dont allow BYOB updates or migrations (#866)
## Description

Just adding a caution to the BYOB doc that we do not allow BYOB updates
or migrations

## Ticket

NA

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.