Note that TensorBoard is only available for SaaS (#869)
## Description

What does the pull request do? If it fixes a bug or resolves a feature
request, be sure to link to that issue.

## Ticket

https://wandb.atlassian.net/browse/DOCS-989

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.