Weave-Models Tutorial (#874)
## Description

Add tutorial for how to integrate W&B Weave with W&B Models. Demo how to
use a fine-tuned Chat Model from the Registry, integrate it into a RAG
Model on Weave, evaluate the new RAG Model using Weave, and save the RAG
Model back to the Registry. Working on this with @knisar.

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.

---------

Co-authored-by: Karan Nisar <knisar17@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Noah Luna <15202580+ngrayluna@users.noreply.github.com>
Co-authored-by: Karan Nisar <karan.nisar@wandb.com>
Co-authored-by: johndmulhausen <john.mulhausen@wandb.com>