Correcting the doc on service accounts to refer to relevant product changes (#884)
## Description

This PR fixes the service account related FAQ doc to accommodate and
refer to the relevant product changes in last few months. This was
missed due to oversight earlier.

## Ticket

NA

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.