Update rate limit section in limits.md with new graphql rate limit (#901)
## Description

What does the pull request do? If it fixes a bug or resolves a feature
request, be sure to link to that issue.

## Ticket

- Update the rate limit section to add new documentation on rate
limiting graphql APIs.
- Fixed some typos
- Reordered parts within the rate limit section
- draft:
https://www.notion.so/wandbai/Public-doc-graphql-rate-limits-12de2f5c7ef3804cb58ef6f675e0dd6e?pvs=4

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [ ] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [ ] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [ ] I merged the latest changes from `main` into my feature branch
before submitting this PR.