Mark projects important with stars, docs (#916)
## Description

Adds a section to describe how to "star" a project. Note that, within
the docs, I refrain from using the feature as a verb. This is done for
readability (e.g. ESL folks)

## Ticket


https://wandb.atlassian.net/browse/DOCS-988?atlOrigin=eyJpIjoiYjJiNTM4NDJmY2E1NDI3ZmJjNGNmZWQ2MDUwNTg0YzEiLCJwIjoiaiJ9

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [ ] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [ ] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [ ] I merged the latest changes from `main` into my feature branch
before submitting this PR.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>