Adding HIPAA retention msg for audit logs (#945)
## Description

Just adding some relevant messaging to the audit logs doc for HIPAA. It
came up as a feedback in a compliance and legal sync.

## Ticket

NA

## Checklist

Check if your PR fulfills the following requirements. Put an `X` in the
boxes that apply.

- [X] Files I edited were previewed on my local development server with
`yarn start`. My changes did not break the local preview.
- [X] Build (`yarn docusaurus build`) was run locally and successfully
without errors or warnings.
- [X] I merged the latest changes from `main` into my feature branch
before submitting this PR.