Remove Qualified pixel (#1118)
Permanently prevents the chat agent from showing up in the docs; action
item derived from chat w/them in which their rep said:

"Got it, thank you. I would recommend we go ahead and remove the
Qualified JS from the docs page. Since it's not a SPA, this should
remove the ability for the experience to "follow" the visitor to the
docs page"