Link to commit in last modified date, remove hash (#1161)
Fixes DOCS-1324.

Removes the commit hash from the "Last modified" footer and makes the
date link to it instead.

I also removed the `td-page-meta` CSS class because it was unused and
repurposed it as the new name for `td-page-meta__lastmod` (which was
misleading, as the div is responsible for more than just the last
modified date).