Minor improvements to Quickstart (#1206)
Mission 1 of this PR: To improve some things that are currently a bit
awkward in the quickstart.
Mission 2: Testing new functionality that deploys CloudFlare preview
builds on PRs created from forks. (See the GitHub Actions
comment/CloudFlare links below)

Mission 2 is why this PR targets the `forks-build-on-cloudflare` branch
(so you can see the CI in action), and is also why the correct merge
order is: this PR first, then #1205 after to finally publish into main.