Add CI that builds CloudFlare Pages previews on forks (#1205)
CloudFlare Pages integration doesn't run on forks. But, because the
forks are public, we can do custom CI that manually `git clone`s those
forks, and, after checking them out, runs the CloudFlare Pages build for
us, commenting back in the originating PR with a link.

See #1206 for an example of this CI in action