Runs: Add beta parameter and argument to code sample (#1331)
Add newly added run setting available to make it easier to restrict
which devices are collecting metrics.

Note: Ideally we'd link to the Reference Docs. However, this feature is
in beta and Settings Class is not currently available in public docs.
For the latter, it'll get updated when we switch to lazydocs.

Jira ticket: https://wandb.atlassian.net/browse/DOCS-1489