Fix table of contents, consolidate fonts (#1374)
1. The Table of Contents has uneven spacing and is very oversized;
CoreWeave sets their TOC `font-size` to a bit below 13px; it's set to
13px here.
2. There is a lot of surface area where the default Docsy font, Open
Sans, pokes through -- even in the main text. Source Sans Pro is used in
other places -- in fact, every time we make an active choice. We
shouldn't whipsaw between two reading fonts, in any case, and Source
Sans Pro seems to be the preference (and, to my eyes, looks more
modern/readable). So, consolidating the reading surfaces to Source Sans
Pro, here.
3. The CoreWeave docs don't need an "On This Page" header above the
Table of Contents and neither do we.

Pay special attention to the TOC on the right in these comparisons, but
also note the more modern-looking font in the main text body:

Before: https://docs.wandb.ai/guides/artifacts/construct-an-artifact/
After:
https://styles.docodile.pages.dev/guides/artifacts/construct-an-artifact/