Create /en subdir in /content for English docs (fixes /support) (#1371)
The Support portal has an issue where the English content being at the
root of /content is scoping all the taxonomy entries from /ko and /ja at
build time and producing a support portal that has English, as well as
Japanese and Korean support articles all rendering together on the
English site. Example: https://docs.wandb.ai/support/anonymous/

In addition we are doing a lot of directory filtering to produce the
right list of stuff when translating the docs, and in general the
content architecture of "ja" and "ko" living among "guides,"
"tutorials," et al is awkward.

This PR fixes all of the above issues. 

**Note: Public-facing URLs are not impacted. This change only impacts
docs contributors.**