Remove global mentions (#1426)
Removes "most" reference to global functions. Examples that were not
modified need a deeper dive so as to not break the code example.

Note: Code snippets, particularly in `integrations/*`, need revisiting
in general. It's unclear how useful they are since they omit key pieces
to make the code useable and/or informative.