Update the HTML preview Github action to follow a newly-added include and list pages that now use it (#1524)
This fixes a situation where a PR adds a new include that is used in one or more existing files. After this fix goes in, the include in the Added section will list and link to all of the pages that use it, and the pages that were modified to use the include will also be listed in the Modified section.