Make CW principal wandb-integration-public for saas (#1569)
Description
-----------
change the principal for multi-tenant saas to the correct principal
`wandb-integration-public` note we changed this recently as
`wandb-integration` will now be used for dedicated customers. There will
be a follow up PR for introducing dedicated cloud usage of principals
but for now to enable saas first going to put out a quicker PR to change
the principal.