Improve efficiency of Cloudflare redirect edge function (#1635)
Improves on the fix deployed in https://github.com/wandb/docs/pull/1631
so that the function fires only when the URL is missing the trailing
slash and would result in a 404. Should reduce the likelihood of hitting
Cloudflare rate limiting, while slightly increasing the likelihood of
getting an obscure 404 if it is not in one of the top level directories
we are considering.