Fix import issues (#1716)
## Description
Fix some one-off issues after the migration

- Fix Models quickstart location
- Fix inadvertent lowercasing of Python API and Weave API refs by the
vendor during the conversion
- Bring in missed Weave reference generation scripts, update them to
preserve the casing in the Weave repo

Verified that all other added, renamed, updated, deleted files in `main`
commits since 9/17 look correct, both for file location and content
changes.

The commit history in this PR is a bit of a mess. It's difficult to do
case-sensitive Git commands on macOS, which has a case-insensitive
filesystem.