Mintlify import (#1727)
## Description
Migrate docs to Mintlify

- Convert content in `wandb/docs` to Mintlify
- Convert `wandb/weave` docs to Mintlify and import them into
`wandb/docs`
- Information architecture updates
- Style updates
- CI/CD updates
  - Port existing Github actions and scripts
- Create Github actions and scripts to support generating `wandb/weave`
reference docs
- TODO finish drafting the description


---
Cut from here down during squash and merge
---

## Potential cut-over scenario
To go live on Mintlify, we would:

1. Sign in to Cloudflare Console and remove the production and PR
deploys. The Cloudflare checks disappear from all existing PRs and
pushing to `main` no longer triggers a Cloudflare deploy.
2. Sign into Mintlify Console and adjust the repo and branch in
https://dashboard.mintlify.com/wb-21fd5541/wb-21fd5541/settings/deployment/git-settings
to point at wandb/docs and the branch where the import is. To check, I
just set it to `wandb/docs` at the `mintlify-import` branch. Also adjust
the site title at the same time. The Mintlify PR checks start running.
3. Just before merging the import branch, sign back into the Mintlify
Console and change the deploy branch to `main`.
4. Merge the import branch to `main`. (_do not_ squash and merge). **The
cut-over is complete.**
5. Remove or archive the `mintlifytest` repo in Github.

> **Notes**:
- The Files Changed and Commits tabs will probably not load in the Git
UI. Check out the branch locally to review it.
- **DO NOT MERGE**. For this test PR, we are importing the Mintlify
content as of the tip of https://github.com/wandb/mintlifytest/pull/42,
which is not yet fully approved and is not yet merged.
- When we do merge an equivalent PR, **we should _not_ squash and
merge**. Otherwise, the commit history of each file will be compressed
into a single commit, defeating the purpose.
    - In this PR, it is expected to that the Cloudflare checks fail.
- While the Mintlify Console is configured to look at this branch in
`wandb/docs`, merging changes in `mintlifytest` will not trigger a
Mintlify deploy. We should switch it back if we want to continue more
iteration in that repo.

## Testing
- [x] Locally, imported content looks right, builds, serves, and
introduces no broken links
- [x] PR tests succeed

---------

Co-authored-by: johndmulhausen <john.mulhausen@wandb.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dan Brian <daniel.brian@gmail.com>
Co-authored-by: Noah Luna <15202580+ngrayluna@users.noreply.github.com>
Co-authored-by: mdlinville <7674613+mdlinville@users.noreply.github.com>