Registry:  Action history v2 (#1831)
## Description

This PR builds off of this pre Mintlify PR:
https://github.com/wandb/docs/pull/1689. Main differences include:

* Moving content from registry lineage page to the more artifact lineage
page
* Added links referencing lineage graphs
* Adding content on how to enable lineage graph tracking
* Word smithing

<!-- Uncomment and add a description of the change -->


<!-- Optionally, uncomment the heading and add details about how you
tested the change and how reviewers should test it. For example:
## Testing
- [ ] Local build succeeds without errors (`mint dev`)
- [ ] Local link check succeeds without errors (`mint broken-links`)
- [ ] PR tests succeed

Replace the `[ ]` with `[x]` to check off the item instead of leaving it
unchecked.

Otherwise, delete this entire section from opening to closing comment.
-->

<!-- Optionally, uncomment the heading and add one or more lines like
these. Otherwise, delete this entire section from opening to closing
comment.

-->

## Related issues

- Fixes https://wandb.atlassian.net/browse/DOCS-1703
- Fixes #12345