[DOCS-1887] Update link checker for Mintlify (#1838)
## Description
Fixes DOCS-1887

In Hugo, the link checker checked files on disk to discover links to
check. This won't work in Mintlify, where content is served from memory
only. This gives us an opportunity to simplify the link checking:

1. Rename existing action and use it to check the live site only. It
will run monthly or on demand and creates a Github issue with the report
only if broken links are found.
2. Add a separate action that will run in a PR. It will run when the PR
branch updates, and will start only after the Mintlify preview
deployment succeeds, similar to how the HTML preview works. It will fail
the PR check if broken links are found and can be made required for a PR
to merge.
3. Updates the .lycheeignore:
    - Consolidate ignored W&B prod URLs using regular expressions
- Ignores URLs that start with `/` or do not start with `http`, which
are internal links checked by `mint broken-links`
    - Reorganizes and categorizes remaining URLs for maintenance
4. Revert the logic I had previously added to try to auto-fix broken
links and cut a PR, because it never worked correctly