Tutorials: Removed PyTorch tutorial (#1872)
## Description

This notebook needs moderate changes. As is, it doesn’t run locally
(missing pip install libraries). It also uses global variables which we
no longer advise. Because of the bad practices, I am removing this
notebook for now. We can add it again once it is up to date.

<!-- Uncomment and add a description of the change -->


<!-- Optionally, uncomment the heading and add details about how you
tested the change and how reviewers should test it. For example:
## Testing
- [ ] Local build succeeds without errors (`mint dev`)
- [ ] Local link check succeeds without errors (`mint broken-links`)
- [ ] PR tests succeed

Replace the `[ ]` with `[x]` to check off the item instead of leaving it
unchecked.

Otherwise, delete this entire section from opening to closing comment.
-->

<!-- Optionally, uncomment the heading and add one or more lines like
these. Otherwise, delete this entire section from opening to closing
comment.
-->

## Related issues

- Fixes
https://wandb.atlassian.net/browse/DOCS-1790?issueKey=DOCS-1790&subProduct=jira-software